Cluster runtime registry and composer controller
